Initial build & Testing environment setup (#1)
Setup of the buildroot project in a separate folder. We override static files in the buildroot folder, this may change at a future stage. 
There is a dockerfile and a docker-compose.yaml, both of which can be used to launch a development container for the project. 
Output of the build is copied to the output folder. 
Initial testing is set up using DejaGnu in the testsuite folder. 

Co-authored-by: Martin Schmidt <martin@scoding.dk>